+++ pkg/querier/blocks_store_balanced_set_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/cortexproject/cortex/pkg/util/services"
+	"github.com/cortexproject/cortex/pkg/util/tls"
 )
 
 func TestBlocksStoreBalancedSet_GetClientsFor(t *testing.T) {
@@ -20,7 +21,7 @@ func TestBlocksStoreBalancedSet_GetClientsFor(t *testing.T) {
 
 	ctx := context.Background()
 	reg := prometheus.NewPedanticRegistry()
-	s := newBlocksStoreBalancedSet(serviceAddrs, log.NewNopLogger(), reg)
+	s := newBlocksStoreBalancedSet(serviceAddrs, tls.ClientConfig{}, log.NewNopLogger(), reg)
 	require.NoError(t, services.StartAndAwaitRunning(ctx, s))
 	defer services.StopAndAwaitTerminated(ctx, s) //nolint:errcheck